org.lockss.plugin.highwire.HighWireArchivalUnit *fixed int/long casting problem
org.lockss.daemon.Configuration
*changed getTimeIntervalParam(string, int) to (string, long)

+++ src/org/lockss/daemon/Configuration.java
@@ -607,7 +607,7 @@ public abstract class Configuration {
   /** Static convenience method to get param from current configuration.
    * Don't accidentally use this on a <code>Configuration</code> instance.
    */
-  public static long getTimeIntervalParam(String key, int dfault) {
+  public static long getTimeIntervalParam(String key, long dfault) {
     return currentConfig.getTimeInterval(key, dfault);
   }
